tests for subworkflow restart and retry
